prevent post deletion without delete code

+++ app/controllers/posts.js
@@ -4,8 +4,15 @@ export default Ember.Controller.extend({
   notNewPost: true,
   actions: {
     deletePost(book) {
-      book.deleteRecord();
-      book.save();
+      var deleteKey = this.get('deleteCode');
+      var modelDeleteCode = book.get('title') + " " + book.get('author');
+      if (deleteKey == modelDeleteCode){
+        book.deleteRecord();
+        book.save();
+        this.set('deleteCode', null);
+      }else{
+        alert('Not authorized to delete this record');
+      }
     },
   },
 });
